Initial code for multi-file link object (contains all Z vectors)
This object maintains the validity of links as they are added or removed and will allow for faster querying of links and enumeration of all linked pairs. It will also prevent the need for validity checks before likelihood and prior calculation b/c it is guaranteed by the add/delete functions.